DynamicSettingsのタイムアウト系フィールドをfloat化する
## Context
check_code_quality.shのpytest実行が17.73秒かかり、PostToolUseフックで
毎回走らせるには重すぎると指摘された。原因はタイムアウト発火を検証する
統合テストで、DynamicSettingsのint制約(gt=0)により最短1秒しか設定でき
ず、実際に1〜2.4秒の実時間待機をしていたこと。

## Decision
asyncioのイベントループ自体を仮想時間化する案は、BaseEventLoop._run_once
という非公開の内部スケジューリングに依存し標準ライブラリの信頼性を犠牲
にするリスクが大きいため見送った。代わりにmax_session_duration_sec/
silence_timeout_sec/mai_timeout_sec/proofread_timeout_sec/
sse_keepalive_sec/intent_translation_timeout_secをintからfloatに変更し、
テストでは0.05秒程度の小さい値を使うようにした。

## Consequences
pytest全体の実行時間が17.73秒→8.86秒に短縮された。JSONC設定ファイルの
数値表現が600から600.0に変わる可能性があるが、json5でのパースや値の
意味には影響しない。

+++ src/shared/config/dynamic_settings.py
@@ -14,25 +14,25 @@ class DynamicSettings(BaseModel):
 
     model_config = ConfigDict(frozen=True)
 
-    max_session_duration_sec: int = Field(default=600, gt=0)
-    silence_timeout_sec: int = Field(default=120, gt=0)
+    max_session_duration_sec: float = Field(default=600, gt=0)
+    silence_timeout_sec: float = Field(default=120, gt=0)
     segment_silence_sec: float = Field(default=3.0, gt=0)
 
     vad_threshold: float = Field(default=0.5, ge=0.0, le=1.0)
 
     mai_locale: str = "ja"
     mai_model_name: str = "mai-transcribe-1"
-    mai_timeout_sec: int = Field(default=60, gt=0)
+    mai_timeout_sec: float = Field(default=60, gt=0)
 
-    proofread_timeout_sec: int = Field(default=15, gt=0)
+    proofread_timeout_sec: float = Field(default=15, gt=0)
     shutdown_delay_sec: float = Field(default=0.5, gt=0)
-    sse_keepalive_sec: int = Field(default=15, gt=0)
+    sse_keepalive_sec: float = Field(default=15, gt=0)
 
     herdr_slot_count: int = Field(default=4, gt=0)
 
     intent_translation_enabled: bool = False
     screenshot_monitor_index: int = Field(default=1, ge=0)
-    intent_translation_timeout_sec: int = Field(default=20, gt=0)
+    intent_translation_timeout_sec: float = Field(default=20, gt=0)
 
     @model_validator(mode="after")
     def _validate_silence_thresholds(self) -> DynamicSettings:
